+++ example/lib/custom_state_object_example.dart
@@ -1,66 +1,87 @@
 import 'package:flutter/material.dart';
 import 'package:upstate/upstate.dart';
 
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+        visualDensity: VisualDensity.adaptivePlatformDensity,
+      ),
+      home: StateWidget(
+        state:MyStateObject(),
+        child: MyHomePage(),
+      ),
+    );
+  }
+}
 
-class CustomStateWidget extends InheritedWidget{
-  final MyState state;
-  CustomStateWidget({@required this.state,Widget child, Key key}):super(key:key);
+class MyHomePage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    var counter = MyStateObject.of(context).counter;
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Flutter Demo Home Page'),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text(
+              'You have pushed the button this many times:',
+            ),
+            CustomStateBuilder(
+                elements: [counter], // will rebuild when any value at this path changes
+                builder: (BuildContext context) {
+                  var state = MyStateObject.of(context);
+                  return Text(
+                    state.counter.count.toString(),
+                    style: Theme.of(context).textTheme.headline4,
+                  );
+                }),
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: counter.increment,
+        tooltip: 'Increment',
+        child: Icon(Icons.add),
+      ),
+    );
+  }
+}
+
+class MyStateObject extends RootStateElement {
+  CounterModel _counter;
+
+  MyStateObject() {
+    _counter = CounterModel(this);
+  }
+
+  CounterModel get counter => _counter;
+
+  static MyStateObject of<T extends StateWidget>(BuildContext context) =>
+      context.dependOnInheritedWidgetOfExactType<T>()?.state;
 
   @override
-  bool updateShouldNotify(CustomStateWidget oldWidget) {
-    if(oldWidget.state!=state){
-      oldWidget.state.unmount();
-      return true;
-    } else{
-      return false;
-    }
+  void unmount() {
+    notifyRemovedFromState();
+    _counter.notifyRemovedFromState();
   }
-}
-
-
-class MyState {
-  final stateObject=StateObject({'counter': Counter()}, typeSafety:TypeSafety.complete, converter:converter);
-  final counterPath = StatePath<CounterModel>(['counter']);
-
-
-  CounterModel get counter => stateObject(counterPath);
-
-  void unmount()=>stateObject.unmount();
-
-  MyState of<T extends CustomStateWidget>(BuildContext context){
-    return context.dependOnInheritedWidgetOfExactType<T>()?.state;
-  }
-}
-
-class Counter {
-  var count = 0;
 }
 
 class CounterModel extends StateElement {
-  final Counter _counter;
+  int _count = 0;
 
-  CounterModel(Counter counter, StateElement parent)
-      : _counter = counter,
-        super(parent);
+  CounterModel(StateElement parent) : super(parent);
 
-  int get count => _counter.count;
+  int get count => _count;
 
   void increment() {
-    _counter.count++;
+    _count++;
     notifyChange();
-  }
-
-  @override
-  toPrimitive() {
-    // TODO: implement toPrimitive
-    throw UnimplementedError();
-  }
-}
-
-StateElement converter(element, parent){
-  if(element is Counter){
-    return CounterModel(element,parent);
-  } else{
-    return null;
   }
 }