+++ lib/upstate.dart
@@ -5,11 +5,12 @@ import 'dart:async';
 import 'src/base.dart';
 export 'src/base.dart';
 
-//TODO: Remove unnecessary child from builder
 
+///A simple inherited widget that is used to hold onto a [StateObject] or your own
+///subclass of [RootStateElement] so that they can be accessed by children.
 class StateWidget extends InheritedWidget {
   final Widget child;
-  final StateObject state;
+  final RootStateElement state;
 
   StateWidget({@required this.state, @required this.child, Key key})
       : super(key: key);
@@ -24,6 +25,8 @@ class StateWidget extends InheritedWidget {
   }
 }
 
+///Provides convenience functions for reducing the amount of boilerplate required
+///for subscribing to changes to [StateElement]s within stateful widgets
 mixin StateConsumerMixin<T extends StatefulWidget> on State<T> {
   List<StreamSubscription> subscriptions = [];
 
@@ -42,14 +45,17 @@ mixin StateConsumerMixin<T extends StatefulWidget> on State<T> {
   }
 }
 
+/// Widget that will will call the builder on changes to any [StateElement] that is specified using [paths]. This
+/// widget will use nearest [StateWidget] parent in the state tree to get its state. If you need a different
+/// state higher up in the tree, you must create an empty subclass of StateWidgeth and then use
+/// StateBuilder<YourStateWidget>.
 class StateBuilder<T extends StateWidget> extends StatefulWidget {
   final List<StatePath> paths;
-  final Widget Function(BuildContext context, StateObject state, Widget child)
+  final Widget Function(BuildContext context, StateObject state)
       builder;
-  final Widget child;
 
   StateBuilder(
-      {@required this.paths, @required this.builder, this.child, Key key})
+      {@required this.paths, @required this.builder, Key key})
       : super(key: key);
 
   @override
@@ -76,6 +82,48 @@ class _StateBuilderState<T extends StateWidget> extends State<StateBuilder>
 
   @override
   Widget build(BuildContext context) {
-    return widget.builder(context, state, widget.child);
+    return widget.builder(context, state);
+  }
+}
+
+/// CustomStateBuilder can be used to rebuild on state changes in multiple states in the widget tree or with a
+/// custom state object that doesn't use [StatePath]s. Since the builder has access to the build context,
+/// you can get whichever state you want or multiple states within the builder. Instead of a list of paths,
+/// CustomStateBuilder takes a list of state elements and rebuilds on changes to any of them. This way you 
+/// can subscribe to elements from different states.
+class CustomStateBuilder extends StatefulWidget {
+  final List<StateElement> elements;
+  final Widget Function(BuildContext context)
+      builder;
+
+  CustomStateBuilder(
+      {@required this.elements, @required this.builder, Key key})
+      : super(key: key);
+
+  @override
+  _CustomStateBuilderState createState() => _CustomStateBuilderState();
+}
+
+class _CustomStateBuilderState extends State<CustomStateBuilder>
+    with StateConsumerMixin {
+ 
+
+  @override
+  void didChangeDependencies() {
+    cancelSubscriptions();
+    for(var element in widget.elements)
+      subscriptions.add(element.subscribe(setStateCallback));
+    super.didChangeDependencies();
+  }
+
+  @override
+  void dispose() {
+    cancelSubscriptions();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return widget.builder(context);
   }
 }